% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_env.R
\name{activate_env}
\alias{activate_env}
\title{Activate conda env}
\usage{
activate_env(conda_env = "echoR", verbose = TRUE)
}
\arguments{
\item{conda_env}{Conda environment to use.}

\item{verbose}{Print messages.}
}
\description{
\link[reticulate]{use_condaenv} often 
\href{https://github.com/rstudio/reticulate/issues/292}{
doesn't work in practice}, especially within RStudio.
\link[echoconda]{activate_env} ensures the env is actually activated 
and used.
}
\examples{
activate_env(conda_env = "echoR")
}
\seealso{
Other conda: 
\code{\link{env_from_yaml}()},
\code{\link{find_env_rlib}()},
\code{\link{install_conda}()}
}
\concept{conda}
