% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_executables_remote.R
\name{check_executables_remote}
\alias{check_executables_remote}
\title{Check executable(s)}
\usage{
check_executables_remote(path, verbose = TRUE)
}
\arguments{
\item{path}{Path or named list of paths to executable(s).}

\item{verbose}{Print messages.}
}
\value{
Logical list.
}
\description{
Check that each executable works. 
Support function for \link[echoconda]{find_executables_remote}.
}
\keyword{internal}
