% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_print.R
\name{cmd_print}
\alias{cmd_print}
\title{Neatly print commands}
\usage{
cmd_print(
  cmd,
  raw = FALSE,
  basepath = TRUE,
  prefix = "...",
  wrap = "\\n   ",
  verbose = TRUE
)
}
\arguments{
\item{cmd}{Command string or list of strings.}

\item{raw}{Print the raw \code{cmd} text.}

\item{basepath}{For each component of \code{cmd} that is a path,
shorten the path to just the basename.}

\item{prefix}{When \code{basepath=TRUE}, append a prefix to the path.}

\item{wrap}{If \code{wrap} is not \code{NULL}, 
use this to separate arguments with newlines.}

\item{verbose}{Print messages.}
}
\description{
Neatly print commands sent to the command line.
}
\examples{
cmd_print("fastqc -d ./fastqc_temp -o ${workdir}
          /fastqFileQC/${sample_name} -f fastq")
}
