% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_list_packages.R
\name{conda_list_packages}
\alias{conda_list_packages}
\title{List conda envs}
\usage{
conda_list_packages(conda_env, conda = "auto", verbose = TRUE)
}
\arguments{
\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{verbose}{Print messages.}
}
\description{
Alternative method for finding packages without relying on the (currently)
internal function \code{reticulate::conda_list_packages}.
}
\keyword{internal}
