% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_echoR_env.R
\name{create_echoR_env}
\alias{create_echoR_env}
\title{Create conda env for \pkg{echolocatoR}}
\usage{
create_echoR_env(
  conda_env = "echoR",
  python_version = NULL,
  channel = c("conda-forge", "bioconda", "r"),
  python_packages = c("pandas>=0.25.0", "pyarrow", "scikit-learn", "bitarray",
    "networkx", "rpy2", "scipy", "pandas-plink"),
  r_packages = c("r-base", "r-devtools"),
  cli_packages = c("tabix", "plink", "macs2"),
  force_install = FALSE,
  auth_token = devtools::github_pat()
)
}
\arguments{
\item{conda_env}{The name of the conda env you want to create.}

\item{python_version}{The version of Python to be used in this Conda
environment. The associated Python package from Conda will be requested
as \code{python={python_version}}. When \code{NULL}, the default \code{python} package
will be used instead. For example, use \code{python_version = "3.6"} to request
that the Conda environment be created with a copy of Python 3.6. This
argument will be ignored if \code{python} is specified as part of the \code{packages}
argument, for backwards compatibility.}

\item{channel}{An optional character vector of Conda channels to include.
When specified, the \code{forge} argument is ignored. If you need to
specify multiple channels, including the Conda Forge, you can use
\verb{c("conda-forge", <other channels>)}.}

\item{python_packages}{Python packages to install.}

\item{r_packages}{R packages to install.}

\item{cli_packages}{Command Line Interface packages to install.}

\item{force_install}{If env already exists, force create a new one.}

\item{auth_token}{GitHub authentication token.}
}
\description{
Create a new env (or update and existing one)
with the necessary Python, R, and command line packages
to run \pkg{echolocatoR}.
}
\details{
By default uses \emph{echoR},
the conda environment distributed with \pkg{echolocatoR}.

\itemize{
\item{plink}{https://anaconda.org/bioconda/plink}
\item{tabix}{https://anaconda.org/bioconda/tabix}
}
}
\seealso{
Other conda: 
\code{\link{activate_env}()},
\code{\link{env_from_yaml}()},
\code{\link{find_env_rlib}()},
\code{\link{install_conda}()}
}
\concept{conda}
