% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env}
\alias{create_env}
\title{Create conda env}
\usage{
create_env(
  conda_env,
  yaml_path,
  method = c("basilisk", "reticulate", "cli"),
  conda = "auto",
  force_new = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{yaml_path}{Path to local or remote yaml file with conda build
specifications.}

\item{method}{Method to use to create the conda env:
\itemize{
\item{"basilisk" : }{Uses the R package \pkg{basilisk}.}
\item{"reticulate" : }{Uses the R package \pkg{reticulate}.}
\item{"cli" : }{Uses a custom wrapper for conda's command line interface.}
}}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{force_new}{If the conda env already exists,
overwrite it with a new one (\emph{DEFAULT}: \code{FALSE}).}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to method-specific 
conda env build functions:
\itemize{
 \item \code{basilisk}: \link[echoconda]{create_env_basilisk}
 \item \code{reticulate}: \link[reticulate]{conda_create}
 \item \code{cli}: \link[echoconda]{create_env_cli} 
 }}
}
\description{
Create conda env using one of several methods.
}
\keyword{internal}
