% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\name{create_env_basilisk}
\alias{create_env_basilisk}
\title{Create conda env: basilisk}
\source{
\href{https://github.com/LTLA/basilisk}{GitHub repo}

\href{https://doi.org/doi:10.18129/B9.bioc.basilisk}{Bioconductor}
}
\usage{
create_env_basilisk(
  yaml_path,
  pkgs,
  sep = "=",
  use_subversion = FALSE,
  use_nodefaults = TRUE,
  start = TRUE
)
}
\arguments{
\item{yaml_path}{Path to local or remote yaml file with conda build
specifications.}

\item{pkgs}{data.table with the columns:
\itemize{
\item{"package" : }{Package names.}
\item{"version" : }{Package versions.}
\item{"channel" : }{Channel where the package was installed from.}
}
This table can be generated by \link[echoconda]{env_to_yaml} function.}

\item{use_subversion}{Specify the specific package subversion.}

\item{start}{Execute \link[basilisk]{basiliskStart} 
to build the actual environment.}
}
\value{
An object of class \link[basilisk]{BasiliskEnvironment}.
}
\description{
Create conda env using one of several methods.
}
\keyword{internal}
