% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env_cli.R
\name{create_env_cli}
\alias{create_env_cli}
\title{Create conda env from yaml: CLI}
\source{
\href{https://stackoverflow.com/questions/35802939/install-only-available-packages-using-conda-install-yes-file-requirements-t}{
StackOverflow suggestions for cross-platform troubleshooting}
}
\usage{
create_env_cli(
  yaml_path = system.file(package = "echoconda", "conda/test.yml"),
  conda = "auto",
  force_new = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{yaml_path}{Path to local or remote yaml file with conda build
specifications.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{force_new}{If the conda env already exists,
overwrite it with a new one (\emph{DEFAULT}: \code{FALSE}).}

\item{verbose}{Print messages.}
}
\description{
Original function for creating conda env from yaml before this was 
implemented in \link[reticulate]{conda_create}.
}
\keyword{internal}
