% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_from_yaml.R
\name{env_from_yaml}
\alias{env_from_yaml}
\title{Create conda env from yaml file}
\usage{
env_from_yaml(
  yaml_path = system.file(package = "echoconda", "conda/echoR.yml"),
  conda = "auto",
  force_new = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{yaml_path}{Path to local or remote yaml file with conda build
specifications.}

\item{conda}{Path to conda executable.}

\item{force_new}{If the conda env already exists,
overwrite it with a new one (\emph{DEFAULT}: \code{FALSE}).}

\item{verbose}{Print messages.}
}
\description{
Create conda env from yaml file
}
\seealso{
Other conda: 
\code{\link{activate_env}()},
\code{\link{find_env_rlib}()},
\code{\link{install_conda}()}
}
\concept{conda}
