% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_conda.R
\name{find_conda}
\alias{find_conda}
\title{Find conda}
\usage{
find_conda(conda = "auto", method = c("basilisk", "reticulate"))
}
\arguments{
\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{method}{Method to use:
\itemize{
\item{"basilisk"}
\item{"reticulate"}
}}
}
\description{
Find conda executable from multiple installations: 
\pkg{reticulate}, \pkg{basilisk}
}
