% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_env_rlib.R
\name{find_env_rlib}
\alias{find_env_rlib}
\title{Find the R library for a specific env}
\usage{
find_env_rlib(conda_env = "echoR", suffix = "lib/R/library/")
}
\arguments{
\item{conda_env}{Conda environment name.}

\item{suffix}{Suffixes to use when searching for R library.}
}
\description{
Find the R library for a specific env
}
\seealso{
Other conda: 
\code{\link{activate_env}()},
\code{\link{env_from_yaml}()},
\code{\link{install_conda}()}
}
\concept{conda}
