% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_executables_remote.R
\name{find_executables_remote}
\alias{find_executables_remote}
\title{Find remotely stored executable(s)}
\usage{
find_executables_remote(
  path = NULL,
  tool = c("bcftools", "gcta", "plink"),
  verbose = TRUE
)
}
\arguments{
\item{path}{User-provided path to executable. 
Is \code{NULL} (default), an executable will be downloaded 
and stored automatically.}

\item{verbose}{Print messages.}
}
\value{
Path to executable.
}
\description{
Find the path to one or more executable binary for each given software.
}
\examples{
paths <- find_executables_remote()
}
\keyword{internal}
