% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_package.R
\name{find_package}
\alias{find_package}
\title{Find package executable}
\usage{
find_package(package, conda_env = "echoR", verbose = TRUE)
}
\arguments{
\item{package}{Name of a package within a conda env.}

\item{conda_env}{Conda environment name.}

\item{verbose}{Print messages.}
}
\description{
Find package executable
}
\examples{
# Tabix
tabix <- find_package(package = "tabix", conda_env = "echoR")
tabix <- find_package(package = "tabix", conda_env = NULL)
# bgzip
bgzip <- find_package(package = "bgzip", conda_env = "echoR")
bgzip <- find_package(package = "bgzip", conda_env = NULL)
}
\seealso{
Other CONDA: 
\code{\link{find_python_path}()}
}
\concept{CONDA}
