% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_conda_basilisk.R
\name{install_conda_basilisk}
\alias{install_conda_basilisk}
\title{Install conda: basilisk}
\usage{
install_conda_basilisk(verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages.}
}
\description{
Install conda if it has not already been installed.
}
\keyword{internal}
