% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_conda_reticulate.R
\name{install_conda_reticulate}
\alias{install_conda_reticulate}
\title{Install conda: reticulate}
\source{
\href{https://github.com/rstudio/reticulate/issues/1055}{
reticulate GitHub Issue.}
\href{https://github.com/conda/conda/issues/5388}{
conda GitHub Issue.}
}
\usage{
install_conda_reticulate(conda, verbose = TRUE, ...)
}
\arguments{
\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments to be passed to 
\link[reticulate]{install_miniconda}.}
}
\description{
Install conda if it has not already been installed.
}
\keyword{internal}
