% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_reticulate.R
\name{install_reticulate}
\alias{install_reticulate}
\title{Install reticulate}
\usage{
install_reticulate(
  dependencies = c("devtools", "reticulate", "lattice", "jsonlite", "Matrix",
    "rappdirs", "Rcpp")
)
}
\description{
\emph{reticulate} often doesn't always install very well via CRAN.
This function helps do it correctly.
}
\keyword{internal}
