% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_envs.R
\name{list_envs}
\alias{list_envs}
\title{List conda envs}
\usage{
list_envs(
  conda = "auto",
  conda_env = NULL,
  method = c("basilisk", "reticulate")
)
}
\arguments{
\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{conda_env}{[Optional] A specific conda environment to search for.}

\item{method}{Method to use:
\itemize{
\item{"basilisk"}
\item{"reticulate"}
}}
}
\description{
List all available conda environments.
}
