% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_packages}
\alias{list_packages}
\title{List conda envs}
\usage{
list_packages(
  packages = NULL,
  conda_env,
  method = c("r", "reticulate", "basilisk"),
  conda = "auto",
  verbose = TRUE
)
}
\arguments{
\item{packages}{A package (e.g. \code{"pandas"})
or a list of packages (e.g. \code{c("pandas","r-dplyr","bgzip")}).
Will recognize R packages with or without the prefixes 
"r-" and "bioconductor-" (i.e. "r-dplyr" is equivalent to "dplyr").}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{method}{Method for listing packages in a given conda env.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{verbose}{Print messages.}
}
\description{
Alternative method for finding packages without relying on the (currently)
internal function \code{reticulate::conda_list_packages}.
}
\keyword{internal}
