% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_yamls.R
\name{list_yamls}
\alias{list_yamls}
\title{List included yaml files}
\usage{
list_yamls(verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages.}
}
\description{
List the yaml files included within \code{echconda}. 
These are provided to make different conda environments that are useful for
specific workflows.
}
\examples{
yamls <- echoconda::list_yamls()
}
