% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_options.R
\name{parse_options}
\alias{parse_options}
\title{Parse options}
\usage{
parse_options(
  path,
  command = NA,
  subcommand = NA,
  help_args = c("-h", "--help", "-help", "-H"),
  args_sep = "\\n",
  args_start = "-+[a-z]+",
  col_names = c("arg", "description"),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to executable file.}

\item{help_args}{Help arguments to use when identifying argument options.}

\item{verbose}{Print messsages.}

\item{sep}{Regex pattern to search for when detecting the beginning 
of each argument description. Can use regex syntax.}
}
\description{
Parse the available options (arguments) for a given command line tool.
}
\examples{
 
opts <- echoconda::parse_options(path="curl")
}
