% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_env.R
\name{remove_env}
\alias{remove_env}
\title{Remove conda environment}
\usage{
remove_env(conda_env, conda = "auto", use_basilisk = FALSE, verbose = TRUE)
}
\arguments{
\item{conda_env}{Name of the conda environment.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{use_basilisk}{Search conda envs created by \pkg{basilisk} 
instead of \pkg{reticulate}.}

\item{verbose}{Print messages.}
}
\description{
Actually remove conda environments,
 unlike \code{reticulate}::\link[reticulate]{conda_remove}.
}
