% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_permissions.R
\name{set_permissions}
\alias{set_permissions}
\title{Set permissions}
\usage{
set_permissions(path, recursive = FALSE, permissions = "rwx", verbose = TRUE)
}
\arguments{
\item{path}{Path to file/folder.}

\item{recursive}{If \code{path} is a folder, 
set permissions set recursively for all files/subfolders within it
 (default: \code{TRUE}).}

\item{permissions}{Permissions to set. 
Defaults to "rwx" (i.e. "read","write" and "execute") for all users.}

\item{verbose}{Print messages.}
}
\description{
Robust methods for setting file/folder permissions 
across multiple OS platforms.
}
\examples{
path <- tempfile()
set_permissions(path=path)
}
