% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rhome.R
\name{set_rhome}
\alias{set_rhome}
\title{Set \code{R_HOME}}
\usage{
set_rhome(conda_env = "echoR_mini", verbose = TRUE)
}
\arguments{
\item{conda_env}{Conda environment to use.}

\item{verbose}{Print messages.}
}
\description{
Set the location of your R library when it is being called through a python 
script (e.g. via the python package \pkg{rpy2}).
}
\keyword{internal}
