% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_env.R
\name{which_env}
\alias{which_env}
\title{Print the currently activated conda env}
\usage{
which_env(verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages.}
}
\description{
\link[reticulate]{py_discover_config}
"This function enables callers to check which versions of
 Python will be discovered on a system as well as which one 
 will be chosen for use with reticulate."
}
