% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_env.R
\name{yaml_to_env}
\alias{yaml_to_env}
\title{Create conda env from yaml file}
\source{
{https://github.com/rstudio/reticulate/issues/779}{GitHub Issue}
}
\usage{
yaml_to_env(
  yaml_path = system.file(package = "echoconda", "conda/echoR.yml"),
  conda = "auto",
  force_new = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{yaml_path}{Path to local or remote yaml file with conda build
specifications.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{force_new}{If the conda env already exists,
overwrite it with a new one (\emph{DEFAULT}: \code{FALSE}).}

\item{verbose}{Print messages.}
}
\description{
Create a conda environment from a yaml file. By default, 
creates the "echoR" conda env to support \pkg{echolocatoR}.
}
\examples{
conda_env <- echoconda::yaml_to_env()
}
\seealso{
Other conda: 
\code{\link{activate_env}()},
\code{\link{env_to_yaml}()},
\code{\link{find_env_rlib}()},
\code{\link{install_conda}()}
}
\concept{conda}
