% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_exists.R
\name{env_exists}
\alias{env_exists}
\title{Check whether a conda env exists}
\usage{
env_exists(env_name, conda = "auto")
}
\arguments{
\item{env_name}{Name of the conda environment.}

\item{conda}{Path to conda executable.}
}
\description{
Check whether a conda env exists
}
\examples{
does_exist <- env_exists(env_name = "echoR")
}
