% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_from_yaml.R
\name{env_from_yaml}
\alias{env_from_yaml}
\title{Create conda env from yaml file}
\usage{
env_from_yaml(
  yaml_path = file.path("https://github.com", "RajLabMSSM/echolocatoR",
    "raw/master/inst", "conda/echoR.yml")
)
}
\arguments{
\item{yaml_path}{Path to local or remote yaml file with conda build
specifications.}
}
\description{
Create conda env from yaml file
}
\seealso{
Other conda: 
\code{\link{activate_env}()},
\code{\link{find_env_rlib}()},
\code{\link{install_conda}()}
}
\concept{conda}
