% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cli.R
\name{import_cli}
\alias{import_cli}
\title{Import Command Line Interface tool}
\usage{
import_cli(
  tool,
  use_namespace = basename(tool),
  conda_env = NULL,
  verbose = TRUE
)
}
\arguments{
\item{use_namespace}{a character string of a namespace to use for the resulting functions}

\item{path}{Path to CLI tool executable file.}
}
\description{
Import a Command Line Interface (CLI) tool into R as a function.
}
\examples{
import_cli(tool="curl")
curl::run("-h")
}
