% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_conda.R
\name{install_conda}
\alias{install_conda}
\title{Install conda if it's missing}
\source{
\href{https://github.com/rstudio/reticulate/issues/1055}{
reticulate GitHub Issue.}
\href{https://github.com/conda/conda/issues/5388}{
conda GitHub Issue.}
}
\usage{
install_conda(conda_path = "auto", verbose = TRUE, ...)
}
\arguments{
\item{conda_path}{Path to conda executable.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments to be passed to 
\link[reticulate]{install_miniconda}.}
}
\description{
Install conda if it's missing
}
\seealso{
Other conda: 
\code{\link{activate_env}()},
\code{\link{env_from_yaml}()},
\code{\link{find_env_rlib}()}
}
\concept{conda}
