% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_conda.R
\name{install_conda}
\alias{install_conda}
\title{Install conda}
\usage{
install_conda(
  method = c("basilisk", "reticulate"),
  conda = "auto",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{method}{Method to use when installing conda.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments to be passed to 
\link[reticulate]{install_miniconda}.}
}
\description{
Install conda if it has not already been installed.
}
\examples{
echoconda::install_conda()
}
