% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_env.R
\name{yaml_to_env}
\alias{yaml_to_env}
\title{Create conda env from yaml file}
\source{
\href{https://github.com/rstudio/reticulate/issues/779}{GitHub Issue}

\href{https://stackoverflow.com/questions/35802939/install-only-available-packages-using-conda-install-yes-file-requirements-t}{
StackOverflow suggestions for cross-platform troubleshooting}
}
\usage{
yaml_to_env(
  yaml_path = system.file(package = "echoconda", "conda", "echoR_mini.yml"),
  method = c("basilisk", "reticulate"),
  conda = "auto",
  force_new = FALSE,
  show_contents = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{yaml_path}{Path to local or remote yaml file with conda build
specifications.}

\item{method}{Method to use to create the conda env:
\itemize{
\item{"basilisk" : }{Uses the R package \pkg{basilisk}.}
\item{"reticulate" : }{Uses the R package \pkg{reticulate}.}
\item{"cli" : }{Uses a custom wrapper for conda's command line interface.}
}}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{force_new}{If the conda env already exists,
overwrite it with a new one (\emph{DEFAULT}: \code{FALSE}).}

\item{show_contents}{Show the contents of the yaml file (if used).}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to method-specific 
conda env build functions:
\itemize{
 \item \code{basilisk}: \link[echoconda]{create_env_basilisk}
 \item \code{reticulate}: \link[reticulate]{conda_create}
 \item \code{cli}: \link[echoconda]{create_env_cli} 
 }}
}
\description{
Create a conda environment from a yaml file. By default, 
creates the "echoR" conda env to support \pkg{echolocatoR}.
}
\examples{
conda_env <- yaml_to_env()
}
\seealso{
Other conda: 
\code{\link{activate_env}()},
\code{\link{env_to_yaml}()},
\code{\link{find_env_rlib}()}
}
\concept{conda}
